moving asset to a types file to be reused in mint and burn
